# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Job Scout API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs with optional filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [new, interested, applied, rejected, ignored, saved, archived]
        - name: source
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: minRelevance
          in: query
          schema:
            type: ["integer", "null"]
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [postedAt, relevanceScore, createdAt, title, company]
            default: createdAt
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: favorite
          in: query
          schema:
            type: ["boolean", "null"]
        - name: employmentType
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobListResponse"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found
    patch:
      operationId: updateJob
      tags: [jobs]
      summary: Update job status, notes, or favorite
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Updated job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /jobs/{id}/summarize:
    post:
      operationId: summarizeJob
      tags: [jobs]
      summary: Re-run AI summarization for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated job with new AI summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"

  # ─── Stats ───────────────────────────────────────────────────────────────────
...
  /keywords:
    get:
      operationId: listKeywords
      tags: [keywords]
      summary: List all search keywords
      responses:
        "200":
          description: List of keywords
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Keyword"
    post:
      operationId: createKeyword
      tags: [keywords]
      summary: Add a new search keyword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordInput"
      responses:
        "201":
          description: Created keyword
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
  /keywords/{id}:
    patch:
      operationId: updateKeyword
      tags: [keywords]
      summary: Update a keyword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KeywordUpdate"
      responses:
        "200":
          description: Updated keyword
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Keyword"
    delete:
      operationId: deleteKeyword
      tags: [keywords]
      summary: Delete a keyword
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Sources ─────────────────────────────────────────────────────────────────
...
  /jobs/{jobId}/notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List notes for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"
    post:
      operationId: createNote
      tags: [notes]
      summary: Add a note to a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Created note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /jobs/{jobId}/notes/{id}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteUpdate"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Social Posts ─────────────────────────────────────────────────────────────
  /jobs/{jobId}/social-post:
    get:
      operationId: getSocialPost
      tags: [social]
      summary: Get generated social post for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Social post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
        "404":
          description: Not generated yet
    post:
      operationId: generateSocialPost
      tags: [social]
      summary: Generate a social post for a job using AI
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SocialPostRequest"
      responses:
        "200":
          description: Generated social post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"

  # ─── Scheduler ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, title, company, source, url, status, relevanceScore, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        source:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [new, interested, applied, rejected, ignored, saved, archived]
        relevanceScore:
          type: ["integer", "null"]
        salary:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        employmentType:
          type: ["string", "null"]
        remote:
          type: ["boolean", "null"]
        postedAt:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        favorite:
          type: boolean
        aiSummary:
          type: ["string", "null"]
        aiRequirements:
          type: ["string", "null"]
        aiWhyFits:
          type: ["string", "null"]
        aiSeniority:
          type: ["string", "null"]
        aiTechnologies:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    JobDetail:
      allOf:
        - $ref: "#/components/schemas/Job"
        - type: object
          properties:
            description:
              type: ["string", "null"]
            notes:
              type: array
              items:
                $ref: "#/components/schemas/Note"
            socialPost:
              oneOf:
                - $ref: "#/components/schemas/SocialPost"
                - type: "null"
    JobUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [new, interested, applied, rejected, ignored, saved, archived]
        favorite:
          type: boolean
    JobListResponse:
      type: object
      required: [jobs, total, page, limit]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Stats ─────────────────────────────────────────────────────────────────
...
    Keyword:
      type: object
      required: [id, term, enabled, createdAt]
      properties:
        id:
          type: integer
        term:
          type: string
        enabled:
          type: boolean
        createdAt:
          type: string
    KeywordInput:
      type: object
      required: [term]
      properties:
        term:
          type: string
          minLength: 1
        enabled:
          type: boolean
    KeywordUpdate:
      type: object
      properties:
        term:
          type: string
          minLength: 1
        enabled:
          type: boolean

    # ── Sources ───────────────────────────────────────────────────────────────
...
    Note:
      type: object
      required: [id, jobId, content, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    NoteInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
    NoteUpdate:
      type: object
      properties:
        content:
          type: string
          minLength: 1

    # ── Social Posts ──────────────────────────────────────────────────────────
    SocialPost:
      type: object
      required: [id, jobId, platform, content, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        platform:
          type: string
          enum: [twitter, linkedin, threads, bluesky]
        content:
          type: string
        tone:
          type: ["string", "null"]
        createdAt:
          type: string
    SocialPostRequest:
      type: object
      required: [platform]
      properties:
        platform:
          type: string
          enum: [twitter, linkedin, threads, bluesky]
        tone:
          type: string
          enum: [applying, interesting, sharing]

    # ── Scheduler ─────────────────────────────────────────────────────────────
...