# lib/api-zod/src/generated/types/jobDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Job Scout API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from './job';
import type { Note } from './note';
import type { SocialPost } from './socialPost';

export type JobDetail = Job & ({
  /** @nullable */
  description?: string | null;
  notes?: Note[];
  socialPost?: SocialPost | null;
});