# lib/api-zod/src/generated/types/jobUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Job Scout API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobUpdateStatus } from './jobUpdateStatus';

export interface JobUpdate {
  status?: JobUpdateStatus;
  favorite?: boolean;
}