# lib/api-zod/src/generated/types/keywordUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Job Scout API specification
 * OpenAPI spec version: 0.1.0
 */

export interface KeywordUpdate {
  /** @minLength 1 */
  term?: string;
  enabled?: boolean;
}