# lib/api-zod/src/generated/types/socialPost.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Job Scout API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SocialPostPlatform } from './socialPostPlatform';

export interface SocialPost {
  id: number;
  jobId: number;
  platform: SocialPostPlatform;
  content: string;
  /** @nullable */
  tone?: string | null;
  createdAt: string;
}